Fix some cases where signal connection IDs where being assigned to guint
Wed Jul 10 14:27:14 2002  Owen Taylor  <otaylor@redhat.com>

        * modules/input/gtkimcontextxim.c gtk/gtkcolorsel.c
        gtk/gtkdialog.c gtk/gtktextbtree.c: Fix some cases
        where signal connection IDs where being assigned to
        guint rather than gulong. (part of #87281, Shivram U)

+++ gtk/gtkcolorsel.c
@@ -136,7 +136,7 @@ struct _ColorSelectionPrivate
   GtkWidget *dropper_grab_widget;
 
   /* Connection to settings */
-  guint settings_connection;
+  gulong settings_connection;
 };
